* structure definitions are now obtained via gccxml; PythonStructs.cs is gone * PythonEnums.cs holds leftover contents of PythonStructs.cs * structs no longer have fields with sepcific enum types; we now have to cast them ourselves * all structs now lack constructors: however, this is easily fixed in python, and rarely a problem in C# * PyComplexObject's apparent layout has changed; some consequential tweaks * a bit less casting in PythonMapper_ files * turned out much easier than expected :)
by: William

+++ src/PythonMapper_strings.cs
@@ -264,7 +264,7 @@ namespace Ironclad
             PyStringObject s = new PyStringObject();
             s.ob_refcnt = 1;
             s.ob_type = this.PyString_Type;
-            s.ob_size = (uint)length;
+            s.ob_size = length;
             s.ob_shash = -1;
             s.ob_sstate = 0;
             Marshal.StructureToPtr(s, data, false);
